Django-Middleware-0x03/requirements.txt
django
djangorestframework
django-filters
djangorestframework-simplejwt
rest-framework-nested
